Backport chapters plugin to old build plugin system
Also fix a number of compile-time warnings, and the inability
to load UI files from the uninstalled directory.

+++ src/plugins/chapters/totem-cmml-parser.c
@@ -134,7 +134,7 @@ totem_cmml_clip_insert_img_attr (TotemCmmlClip	*clip,
 	GdkPixdata	*pixdata;
 	GdkPixbuf	*pixbuf;
 	guchar		*base64_dec;
-	guint		st_len;
+	gsize		st_len;
 	GError		*error = NULL;
 
 	g_return_if_fail (clip != NULL);
